requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
pinecone==5.3.1
langchain-groq
transformers